+++ src/views/Example/ChildComponent.js
