+++ src/views/Example/MyComponent.js
@@ -1,5 +1,5 @@
 import React from "react";
-import ChildComponent from "../ChildComponent";
+import ChildComponent from "./ChildComponent";
 class MyComponent extends React.Component {
     state = {
         firstName: '',
